Navbar Tests
