+++ src/main/resources/templates/forgotPassword.html
@@ -5,7 +5,7 @@
     <title>Acebook Forgot Password Page</title>
 </head>
 <body>
-<div class="topnav">
+<div class="navbar">
     <a class="active"  href="/login">Sign In</a>
 </div>
