+++ src/main/resources/templates/homepage.html
@@ -5,9 +5,9 @@
   <title>Acebook Home Page</title>
 </head>
 <body>
-<div class="topnav">
-  <a class="active"  href="/login">Sign In</a>
-  <a href="/users/new">Sign Up</a>
+<div class="navbar">
+  <a class="active"  href="/login" id="login">Sign In</a>
+  <a href="/users/new" id="signup">Sign Up</a>
 <!--  <a href="/posts">Posts</a>-->
 <!--  <a href="/logout">Log Out</a>-->
 </div>
