+++ src/main/resources/templates/login.html
@@ -5,8 +5,8 @@
     <title>Please sign in</title>
 </head>
 <body>
-    <div class="topnav">
-        <a class="active" href="/homepage">Home</a>
+    <div class="navbar">
+        <a class="active" href="/homepage" id="home">Home</a>
 <!--        <a href="/posts">Posts</a>-->
 <!--        <a href="/logout">Log Out</a>-->
     </div>
