+++ src/main/resources/templates/logout.html
@@ -3,14 +3,14 @@
         xmlns:sec="https://www.thymeleaf.org/thymeleaf-extras-springsecurity5" >
 <head>
     <meta charset="UTF-8">
-    <title>Logout Page</title>
+    <title>Logout</title>
     <link rel="stylesheet" href="/main.css" />
 </head>
 <body>
-<div class="topnav">
+<div class="navbar">
 <!--    <a href="/homepage">Home</a>-->
-    <a href="/posts">Posts</a>
-    <a th:href="${user}">Profile</a>
+    <a href="/posts" id="posts">Posts</a>
+    <a th:href="${user}" id="profile">Profile</a>
     <a class="active" href="/logout">Log Out</a>
     <div sec:authorize="isAuthenticated()" style="text-align: right; padding: 5px; color: white; padding-top: 15px; padding-right: 30px;">
         Signed in as: <span sec:authentication="name"></span>
