+++ src/main/resources/templates/users/new.html
@@ -7,8 +7,8 @@
         <link rel="stylesheet" href="/main.css" th:href="@{/main.css}" />
     </head>
     <body>
-    <div class="topnav">
-        <a href="/homepage">Home</a>
+    <div class="navbar">
+        <a href="/homepage" id="home">Home</a>
     </div>
         <h1>Sign Up to Acebook</h1>
         <p style="text-align:center">Please fill in this form to create an account.</p>
