+++ src/test/java/HomePageTest.java
@@ -8,42 +8,36 @@ import org.junit.runner.RunWith;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringBootTest(classes = Application.class)
-public class SignUpTest {
 
+public class HomePageTest {
     WebDriver driver;
     Faker faker;
-
     @Before
     public void setup() {
         System.setProperty("webdriver.chrome.driver", "/usr/local/bin/chromedriver");
         driver = new ChromeDriver();
         faker = new Faker();
-        driver.get("http://localhost:8080/users/new");
+        driver.get("http://localhost:8080/homepage");
     }
-
     @After
     public void tearDown() {
         driver.close();
     }
-
     @Test
-    public void successfulSignUpRedirectsToSignIn() {
-        driver.findElement(By.id("username")).sendKeys(faker.internet().emailAddress());
-        driver.findElement(By.id("password")).sendKeys("Password098!");
-        driver.findElement(By.id("submit")).click();
+    public void successfulRedirectToSignUpPage() {
+        driver.findElement(By.id("signup")).click();
         String title = driver.getTitle();
-        Assert.assertEquals("Please sign in", title);
+        Assert.assertEquals("Signup", title);
     }
     @Test
-    public void successfulRedirectToHomePage() {
-        driver.findElement(By.id("home")).click();
+    public void successfulRedirectToLoginPage() {
+        driver.findElement(By.id("login")).click();
         String title = driver.getTitle();
-        Assert.assertEquals("Acebook Home Page", title);
+        Assert.assertEquals("Please sign in", title);
     }
 }
