+++ src/test/java/ProfileTest.java
@@ -1,4 +1,5 @@
 import com.github.javafaker.Faker;
+import com.github.javafaker.IdNumber;
 import com.makersacademy.acebook.Application;
 import org.junit.After;
 import org.junit.Assert;
@@ -13,10 +14,9 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringBootTest(classes = Application.class)
-public class LogoutTest {
+public class ProfileTest {
     WebDriver driver;
     Faker faker;
-
     @Before
     public void setup() {
         System.setProperty("webdriver.chrome.driver", "/usr/local/bin/chromedriver");
@@ -31,24 +31,16 @@ public class LogoutTest {
         driver.findElement(By.id("username")).sendKeys(username);
         driver.findElement(By.id("password")).sendKeys("Password098!");
         driver.findElement(By.id("submit")).click();
-        driver.get("http://localhost:8080/logout");
+        driver.findElement(By.id("content")).sendKeys("LikeTest");
+        driver.findElement(By.id("submit")).click();
+        driver.findElement(By.id("like")).click();
+        driver.findElement(By.id("profile")).click();
     }
     @After
     public void tearDown() {
         driver.close();
     }
-    @Test
-    public void successfulLogoutRedirectToLoginPage() {
-        driver.findElement(By.id("logout")).click();
-        String title = driver.getTitle();
-        Assert.assertEquals("Acebook Home Page", title);
-    }
-    @Test
-    public void successfulLogoutRedirectToPostsPageOnCancel() {
-        driver.findElement(By.id("cancel")).click();
-        String title = driver.getTitle();
-        Assert.assertEquals("Acebook", title);
-    }
+
     @Test
     public void successfulRedirectToPostsPage() {
         driver.findElement(By.id("posts")).click();
@@ -56,9 +48,17 @@ public class LogoutTest {
         Assert.assertEquals("Acebook", title);
     }
     @Test
-    public void successfulRedirectToProfilePage() {
-        driver.findElement(By.id("profile")).click();
+    public void successfulRedirectToLogoutPage() {
+        driver.findElement(By.id("logout")).click();
         String title = driver.getTitle();
-        Assert.assertEquals("Profile", title);
+        Assert.assertEquals("Logout", title);
+    }
+    @Test
+    public void LikesIncrease(){
+        String likes = String.valueOf(driver.findElement(By.id("NoLikes")).getText());
+        Assert.assertEquals("1", likes);
+        driver.findElement(By.id("like")).click();
+        String likes2 = String.valueOf(driver.findElement(By.id("NoLikes")).getText());
+        Assert.assertEquals("2", likes2);
     }
 }
